+++ Cargo.toml
@@ -22,14 +22,18 @@ categories = ["api-bindings", "filesystem"]
 maintenance = { status = "actively-developed" }
 travis-ci   = { repository = "inotify-rs/inotify" }
 
+[features]
+default = ["stream"]
+stream = ["futures", "mio", "tokio-io", "tokio-reactor"]
+
 [dependencies]
 bitflags      = "1"
-futures       = "0.1"
+futures       = { version = "0.1", optional = true }
 inotify-sys   = "0.1.1"
 libc          = "0.2"
-mio           = "0.6"
-tokio-io      = "0.1"
-tokio-reactor = "0.1"
+mio           = { version = "0.6", optional = true }
+tokio-io      = { version = "0.1", optional = true }
+tokio-reactor = { version = "0.1", optional = true }
 
 [dev-dependencies]
 tempdir = "0.3"
